// src/types/items.ts
export declare interface IPersonal {
  name: string
  email: string
  phone: string
}

export declare interface IStep2 {
  id: string
  name: string
  monthly: string
  yearly: string
  discount: string
  icon: string
}

export declare interface IStep3 {
  id: string
  title: string
  monthly: string
  yearly: string
  semititle: string
}